Don't display checkboxes for single member role groups
"Removing" a user from such a group merely replaces them with the nobody
user.  As such, providing the illusion that one can remove users from
the role is confusing.

If a right existed to limit the set of users who would be owners of an
asset, displaying it as a drop-down (as is done for Tickets) would be
sensible.