Summarize linked assets on the ticket display page
Linked assets are summarized in an accordion much like requestors are
summarized in "More About Requestors".  We use the accordion rather than
a collection list since there are likely to be only a few assets linked
per ticket and little need to easily compare many rows.  An accordion
provides a better experience and more flexibility in extending the asset
summary.