Allow passing of id during RT::Asset->Create
When importing from an external source, it may be useful to allow
creation of assets with specifically-crafted ids, soas to match RT's
concept of "asset id" with the external source's.

Note that this does not deal with increasing the Asset id sequence to
above the newly created asset's id.  Failure to do so will result in
errors when the sequence reaches an id which is already in use.  While
mysql and sqlite will adjust their auto_increment, postgres and Oracle
will not adjust their sequences.

As id is not included in RT::Asset->WritableAttributes, this does not
allow creation of assets with arbitrary ids from the web UI, even if an
"id" parameter is explicitly submitted.