don't position "User Summary" as "absolute right" for small screens.
by default(in RT's accordion.css), position css of "User Summary" link(used in
ShowRoleMembers) is:

{
    position: absolute;
    right: 0;
}

this could result in overlapping(with the content at the left side) when the
screen is not wide enough. I tried "float: right", but then it shadows the
following contents(like "Real Name Enoch Root") below it in small screens.

using "relative" position in small screens works(no overlapping or hidding),
though it doesn't look that good(since it's not at the right side).