Tighten up labels to reduce wrapping problems with inline-block
Redering <td>s as inline-block brings in somewwhat confusing redering
when the width is insufficient for the elements, wrapping them to the
next line.  Reduce the size of the labels to make the range of widths
with these problems smaller; once the window width is small enough, the
rendering switches to single-column, with none of these issues.