Stop forcing nowrap onto ticket searches
When 7a0a4573 said that we might want this for core RT it forgot that RT
doesn't cap the width of search boxes.  This makes tickets with long
subjects push the More about requestors box ticket list to absurd
widths.  It breaks portlets on the homepage.  It pushes Dashboard ticket
lists to no longer fit on the dashboard page, especially if you have
anything in the sidebar.

Anything past about 50/60 characters will show the behavior, for
excessive hilarity push it to 100 characters.

This may be binding this change too tightly to the search results.
There may be other Asset UI where we wanted to disable wrapping, but the
original commit didn't give enough clues.