Allow the standard install order by temporarily enabling the plugin
Assets required that it be placed in @Plugins prior to "make initdb"
because it required that the 'assets' lifecycle (from Assets_Config.pm)
be available during initialdata insert.

Alllow the more standard "make install initdb" followed by enabling the
plugin, by enabling the plugin in the initialdata if it is not already
enabled.