Support the NULL value for Custom Fields
/Elements/SelectCustomFieldValue which is used on the Asset Search page
presents -, (no value), rest of values for Select type CFs.

(no value) is NULL in the backend because Query Builder turns that into
CF.{foo} IS NOT NULL.

Assets has a much simpler search and was only really supporting the -
(unpicked) or value version.  It's simple to case on NULL as an argument
to do a join out and generate the IS NOT NULL, rather than the query
which was generated which was Content not like '%NULL%' which might
match, but never in the intended manner.