refactor to not convert OrderBy cfs to objects
the OrderBy arg is used in /Elements/CollectionAsTable/Header to determine if
to flip Order or not, converting it to object breaks it.

the reason of conversion was to make sure we load the correct OrderBy cf
(since we have Catalog info in ProcessAssetsSearchArguments but not so in
RT::Assets::OrderByCols), to refactor it, I added Catalog id to RT::Assets
object hash when limiting Catalog so OrderByCols can make use of it.