Always return undef from _Value on permission denied
The bare return will produce an empty list when called in list context.  This
produced "Odd number of elements in hash assignment" warnings when ->Catalog
was used in the parameter list for LoadByNameAndCatalog inside
LoadCustomFieldByIdentifier.

DBIx::SearchBuilder::Record's _Value (by way of __Value) makes sure to always
return a scalar.